fix(security): Phase 5a + Phase 1d — Poseidon domain constants + ChainWork u128
## Phase 5a (H11): Poseidon domain separation constants
- DRK_POSEIDON_DOMAIN_NULLIFIER, _TOKEN_COMMIT, _TX_BINDING, _COIN_COMMIT, _MERKLE_LEAF
- Distinct pallas::Base values prepended to poseidon_hash calls
- Follows existing DRK_* naming convention (HAZOP naming refinement)
- Closes: H11 (no domain separation in Poseidon instances)

## Phase 1d (M1): ChainWork AtomicU64 → Mutex<u128>
- Eliminates u64 overflow at ~4.3B blocks (~16,350 years at 120s/block)
- Mutex contention negligible: writes once per block, reads on stratum/RPC
- Sled persistence updated: 16-byte u128 LE encoding (clean break)
- Closes: M1 (chain work wraps at u64 boundary)

+++ src/sdk/src/blockchain.rs
@@ -335,10 +335,10 @@ impl BlockTarget {
     /// Chain work contributed by a block at this target.
     /// `work = u32::MAX / target`. Higher = heavier chain = fork selection prefers.
     /// Returns 0 for zero target (degenerate — never valid).
-    /// G11: encapsulates `u32::MAX as u64 / target as u64`.
-    pub const fn chain_work(self) -> u64 {
+    /// Returns u128 per ChainWork migration (M1) — u64 could wrap at ~4.3B blocks.
+    pub const fn chain_work(self) -> u128 {
         if self.0 == 0 { return 0; }
-        u32::MAX as u64 / self.0 as u64 // G11: u32→u64 is infallible widening
+        u32::MAX as u128 / self.0 as u128
     }
 
     /// Conventional mining difficulty for display/pool purposes.
